Extract class CommandMenuProvider from class MainMenu.
