+++ Dev/Editor/Effekseer/GUI/Menu/MainMenu.cs
@@ -144,211 +144,11 @@ namespace Effekseer.GUI.Menu
 
 		private void ReloadMenu()
 		{
-			this.Controls.Add(SetupFilesMenu(recentFilesMenuManager));
-			this.Controls.Add(SetupEditMenu());
-			this.Controls.Add(SetupViewMenu());
+			this.Controls.Add(CommandMenuProvider.SetupFilesMenu(recentFilesMenuManager));
+			this.Controls.Add(CommandMenuProvider.SetupEditMenu());
+			this.Controls.Add(CommandMenuProvider.SetupViewMenu());
 			this.Controls.Add(WindowMenu.SetupWindowMenu());
-			this.Controls.Add(SetupHelpMenu());
-		}
-
-		private Menu SetupHelpMenu()
-		{
-			var menu = new Menu(new MultiLanguageString("Help"));
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.ViewHelp));
-			//menu.Controls.Add(create_menu_item_from_commands(Commands.OpenSample));
-
-			menu.Controls.Add(new MenuSeparator());
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.About));
-
-			return menu;
-		}
-
-		private Menu SetupViewMenu()
-		{
-			var view = new MultiLanguageString("View");
-
-			var menu = new Menu(view);
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Play));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Stop));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Step));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.BackStep));
-
-			return menu;
-		}
-
-		private Menu SetupEditMenu()
-		{
-			var edit = new MultiLanguageString("Edit");
-
-			var menu = new Menu(edit);
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.AddNode));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.InsertNode));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.RemoveNode));
-
-			menu.Controls.Add(new MenuSeparator());
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Copy));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Paste));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.PasteInfo));
-
-			menu.Controls.Add(new MenuSeparator());
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Undo));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Redo));
-
-			return menu;
-		}
-
-		private static Menu SetupFilesMenu(RecentFilesMenuManager recentFiles)
-		{
-			var file = new MultiLanguageString("Files");
-			var input = new MultiLanguageString("Import");
-			var output = new MultiLanguageString("Export");
-
-			var menu = new Menu(file);
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.New));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Open));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Overwrite));
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.SaveAs));
-
-			menu.Controls.Add(new MenuSeparator());
-
-			menu.Controls.Add(SetupImportSubMenu(input));
-
-			menu.Controls.Add(SetupExportSubMenu(output));
-
-			menu.Controls.Add(new MenuSeparator());
-
-			{
-				recentFiles.Reload();
-				menu.Controls.Add(recentFiles.Menu);
-			}
-
-			menu.Controls.Add(new MenuSeparator());
-
-			menu.Controls.Add(CreateMenuItemFromCommands(Commands.Exit));
-
-			return menu;
-		}
-		
-
-		private static MenuItem CreateMenuItemFromCommands(Func<bool> onClicked)
-		{
-			var item = new MenuItem();
-			var attributes = onClicked.Method.GetCustomAttributes(false);
-			var uniquename = UniqueNameAttribute.GetUniqueName(attributes);
-			item.Label = NameAttribute.GetName(attributes);
-			item.Shortcut = Shortcuts.GetShortcutText(uniquename);
-			item.Clicked += () => { onClicked(); };
-
-			return item;
-		}
-
-		private static Menu SetupExportSubMenu(MultiLanguageString output)
-		{
-			var export_menu = new Menu(output, Icons.Empty);
-
-			for (int c = 0; c < Core.ExportScripts.Count; c++)
-			{
-				var item = new MenuItem();
-				var script = Core.ExportScripts[c];
-				item.Label = script.Title;
-				item.Clicked += () =>
-				{
-					var filter = script.Filter.Split('.').Last();
-					var result = swig.FileDialog.SaveDialog(filter, System.IO.Directory.GetCurrentDirectory());
-
-					if (!string.IsNullOrEmpty(result))
-					{
-						if (System.IO.Path.GetExtension(result) != "." + filter)
-						{
-							result += "." + filter;
-						}
-
-						var filepath = result;
-						script.Function(filepath);
-
-						System.IO.Directory.SetCurrentDirectory(System.IO.Path.GetDirectoryName(filepath));
-					}
-
-					/*
-							SaveFileDialog ofd = new SaveFileDialog();
-
-							ofd.InitialDirectory = System.IO.Directory.GetCurrentDirectory();
-							ofd.Filter = script.Filter;
-							ofd.FilterIndex = 2;
-							ofd.OverwritePrompt = true;
-
-							if (ofd.ShowDialog() == DialogResult.OK)
-							{
-								var filepath = ofd.FileName;
-								script.Function(filepath);
-
-								System.IO.Directory.SetCurrentDirectory(System.IO.Path.GetDirectoryName(filepath));
-							}
-							else
-							{
-								return;
-							}
-							*/
-				};
-				export_menu.Controls.Add(item);
-			}
-
-			return export_menu;
-		}
-
-		private static Menu SetupImportSubMenu(MultiLanguageString input)
-		{
-			var import_menu = new Menu(input, Icons.Empty);
-
-			for (int c = 0; c < Core.ImportScripts.Count; c++)
-			{
-				var item = new MenuItem();
-				var script = Core.ImportScripts[c];
-				item.Label = script.Title;
-				item.Clicked += () =>
-				{
-					var filter = script.Filter.Split('.').Last();
-					var result = swig.FileDialog.OpenDialog(filter, System.IO.Directory.GetCurrentDirectory());
-
-					if (!string.IsNullOrEmpty(result))
-					{
-						var filepath = result;
-
-						script.Function(filepath);
-						System.IO.Directory.SetCurrentDirectory(System.IO.Path.GetDirectoryName(filepath));
-					}
-
-					/*
-							OpenFileDialog ofd = new OpenFileDialog();
-
-							ofd.InitialDirectory = System.IO.Directory.GetCurrentDirectory();
-							ofd.Filter = script.Filter;
-							ofd.FilterIndex = 2;
-							ofd.Multiselect = false;
-
-							if (ofd.ShowDialog() == DialogResult.OK)
-							{
-								var filepath = ofd.FileName;
-								script.Function(filepath);
-
-								System.IO.Directory.SetCurrentDirectory(System.IO.Path.GetDirectoryName(filepath));
-							}
-							else
-							{
-								return;
-							}
-							*/
-				};
-				import_menu.Controls.Add(item);
-			}
-
-			return import_menu;
+			this.Controls.Add(CommandMenuProvider.SetupHelpMenu());
 		}
 
 		private void MainForm_Load(object sender, EventArgs e)
